+++ lib/generators/cms/component/contact_page/contact_page_generator.rb
@@ -19,6 +19,16 @@ module Cms
               model.title = 'Page: Contact'
               model.attributes = [
                 {
+                  name: 'headline',
+                  type: :string,
+                  title: 'Headline',
+                },
+                {
+                  name: 'content',
+                  type: :html,
+                  title: 'Content',
+                },
+                {
                   name: crm_activity_type_attribute_name,
                   type: :string,
                   title: 'CRM Activity Type',