+++ lib/generators/cms/kickstart/templates/app/models/container.rb
@@ -1,4 +1,5 @@
 class Container < Obj
+  include Cms::Attributes::Headline
   include Cms::Attributes::ShowInNavigation
 
   def page