+++ lib/generators/cms/widget/video/templates/app/models/video_widget.rb
@@ -1,6 +1,8 @@
 class VideoWidget < Obj
   include Widget
 
+  include Cms::Attributes::Headline
+  include Cms::Attributes::Content
   include Cms::Attributes::Source
   include Cms::Attributes::Width
   include Cms::Attributes::Height