+++ spec/generators/widget/slider_generator_spec.rb
@@ -39,7 +39,7 @@ describe Cms::Generators::Widget::SliderGenerator do
               file 'indicators.html.haml'
               file 'left_control.html.haml'
               file 'right_control.html.haml'
-              file 'title.html.haml'
+              file 'headline.html.haml'
             end
           end
         end
@@ -48,14 +48,17 @@ describe Cms::Generators::Widget::SliderGenerator do
           directory 'cms' do
             directory 'attributes' do
               file 'images.rb'
+              file 'headline.rb'
             end
           end
         end
 
         directory 'widgets' do
           directory 'slider_widget' do
-            file 'show.html.haml'
-            file 'thumbnail.html.haml'
+            directory 'views' do
+              file 'show.html.haml'
+              file 'thumbnail.html.haml'
+            end
 
             directory 'locales' do
               file 'de.slider_widget.yml'